backend/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
reportlab==4.2.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
